requirements.txt
panda3d
colorama
panda3d-gltf
panda3d-simplepbr
